Fixed issue where certain values would not output
When running the code with various json files, certain data would not output into the table at all. I figured out that the two if statements that check if the data is unicode or int/float is not enough. For some reason, some data is not an instance of any of these and therefore will never output. So, unless the entry is a list or a dict, I returned the string version of entry. It works perfectly!